+++ angular-workspace/projects/trimble-oss/modus-angular-components/src/lib/stencil-generated/index.ts
@@ -12,6 +12,8 @@ export const DIRECTIVES = [
   d.ModusCheckbox,
   d.ModusChip,
   d.ModusDataTable,
+  d.ModusDateInput,
+  d.ModusDatePicker,
   d.ModusDropdown,
   d.ModusFileDropzone,
   d.ModusList,